requirements.txt
PySimpleGUI==5.0.6
requests==2.32.3
alive_progress==3.1.5
gitpython==3.1.43